requirements.txt
flywheel-gear-toolkit==0.6.14
flywheel-gears==0.3.0
flywheel-sdk==16.19.2
fw-core-client==1.1.2
nibabel==3.2.2
numpy==1.24.0
